# src/main/java/com/rca/agent/chat/ChatController.java
package com.rca.agent.chat;

import jakarta.validation.Valid;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;

@RestController
@RequestMapping("/api/v1/rca/chat")
@Validated
@CrossOrigin(origins = "*")
public class ChatController {

  private final ChatService chatService;

  public ChatController(ChatService chatService) {
    this.chatService = chatService;
  }

  @PostMapping
  public ResponseEntity<ChatResponse> chat(@Valid @RequestBody ChatRequest request) {
    ChatResponse response = chatService.chat(request);
    return ResponseEntity.ok(response);
  }
}